+++ NEWS.md
@@ -1,8 +1,9 @@
-# unimod 0.x
+# PTMods 0.x
 
-## unimod 0.3 
+## PTMods 0.3
 
-- Add `convertAnnotation()` to parse and convert between 3 PTM annotation 
+- Rename the package to `PTMods`.
+- Add `convertAnnotation()` to parse and convert between 3 PTM annotation
   styles. Currently only supports `unimodId`, unimod `name` and `deltaMass`.
 
 ## unimod 0.2 [2023-02-05]